Separated `frame_with_value` from `spawn.h`.
